backend/requirements.txt
# /backend/requirements.txt

Flask==2.1.2
python-dotenv==0.21.0
openai==0.27.0
requests==2.28.1
psycopg2==2.9.3
SQLAlchemy==1.4.36
alembic==1.7.7
redis==4.3.3
celery==5.2.7
boto3==1.26.50
stripe==4.0.0
PyJWT==2.4.0
prometheus_client==0.15.0
pytest==7.1.2
pydantic==1.10.2
flasgger==0.9.5
flask-limiter==2.6.0
flask-talisman==0.8.1
